# nodejsSwagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Benim Uygulamam",
    "description": "Bu uygulama hakkında bilgiler",
    "termsOfService": "test",
    "contact": {
      "email": "kkbra.celik92@gmail.com"
    },
    "license": {
      "name": "test",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.11"
  },
  "externalDocs": {
    "description": "deneme",
    "url": "http://swagger.io"
  },
  "servers": [
    {
      "url": "http://localhost:5000/"
    }
  ],
  "tags": [
    {
      "name": "Team",
      "description": "Kullanıcı Hakkında Kullanılabilecek Endpointler"
    }
  ],
  "paths": {
    "/create": {
      "post": {
        "tags": ["Team"],
        "summary": "Takım Oluştur",
        "description": "Burada Takım Oluşturulur",
        "requestBody": {
          "description": "Bir Takım Oluşturulur",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Team"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Team"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Team"
                }
              }
            }
          }
        }
      }
    },
    "/list": {
      "get": {
        "tags": ["Team"],
        "summary": "Mevcut Takım Kayıtları",
        "description": "Takımlar",
        "parameters": [
          {
            "name": "teamId",
            "in": "query",
            "description": "Belirli bir takımı bulmak için",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Başarılı Yanıt",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamList"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Geçerli Parametreler Verilmedi"
          }
        }
      }
    },
    "/delete/{id}": {
      "delete": {
        "tags": ["Team"],
        "summary": "Takım Sil",
        "description": "Takım Siler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Silinecek Kayıt Numarası",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Başarılı Yanıt",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/DeleteSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Geçersiz Numara"
          },
          "404": {
            "description": "Böyle Bir Takım Mevcut Değil"
          }
        }
      }
    },
    "/update/{id}": {
      "put": {
        "tags": ["Team"],
        "summary": "Takım Günceller",
        "description": "Takım Güncelle",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Güncellenecek Kayıt Numarası",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Güncellenecek Kayıt Objesi",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeam"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeam"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Başarılı Yanıt",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/UpdateModel"
                }
              }
            }
          },
          "400": {
            "description": "Geçersiz Numara"
          },
          "404": {
            "description": "Böyle Bir Takım Mevcut Değil"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Team": {
        "type": "object",
        "properties": {
          "team": {
            "type": "string",
            "example": "theUser"
          }
        }
      },
      "TeamList": {
        "type": "object",
        "properties": {
          "team": {
            "type": "string",
            "example": "theUser"
          },
          "id": {
            "type": "number"
          }
        }
      },
      "DeleteSuccess": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "theUser"
          }
        }
      },
      "UpdateTeam": {
        "type": "object",
        "properties": {
          "team": {
            "type": "string",
            "example": "theUser"
          }
        }
      },
      "UpdateModel": {
        "type": "object",
        "properties": {
          "team": {
            "type": "string",
            "example": "theUser"
          },
          "id": {
            "type": "integer",
            "example": "theUser"
          }
        }
      },
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "requestBodies": {
      "Pet": {
        "description": "Pet object that needs to be added to the store",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Pet"
            }
          },
          "application/xml": {
            "schema": {
              "$ref": "#/components/schemas/Pet"
            }
          }
        }
      },
      "UserArray": {
        "description": "List of user object",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        }
      }
    }
  }
}